Partly removed need for .finish() on tcp.*
Created system that allows tcp.Direction to figure out as
much about seq_start and final arrival times as possible with
the packets that have arrived. includes making Direction.data
and Direction.seq_start into properties that are re-calculated
on the fly, based on the packets available and whether there's
a handshake.