Be smarter about message body presence/absence
Be sure not to process bodies for 204 and 304 responses,
and don't bother handling pipelining for HTTP/1.0 without
a content-length. (Strictly, 1.0 should never pipeline, but
IIS at least does this, but it gives a content-length so
we handle it just fine). This fix results in more accurate
processing of out-of-order.pcap, so the test har was updated.
There's also a couple useful debugging classes in pcaputil.