Fix tcp.Direction assertion failure
Remove useless asserts from packet-bridging-the-gap scenario. It
turns out to be possible for the second chunk to merge onto the
new one without adding any new data, if it's completely inside
the newly merged chunk in front of it. The other condition is
impossible because of the sorting invariant (I think).

Also, somewhere in the merging, clear_data didn't get the news
that self.chunks is a SortedCollection and you can't del[] it.