Misc TLS fixes: -d, --tls, and others
Make sure TLS works with -d. Make it optional, with --[no-]tls flags.
Cipher suite is now left as int by dpkt instead of CipherSuite obj, so
we have to adjust to that. Also logging, comments, and stuff.
fhs_ssl.pcap works when you manually supply the keylog, but the test
script doesn't, so it looks broken.